refactor: 🔀 move generate-otp to utils, add to index.js, and rename util files to kebab-case
