+++ server/utils/index.js
@@ -3,13 +3,15 @@ const { verifiedUserTemplate } = require("./custom_templates");
 const {
   generateAccessToken,
   generateRefreshToken,
-} = require("./generateTokens");
+} = require("./generate-tokens");
+const genratedOTP = require("./generate-otp");
 
 module.exports = {
   ...constants,
   templates: {
     verifiedUserTemplate,
   },
+  genratedOTP,
   tokens: {
     generateAccessToken,
     generateRefreshToken,
